Update to elemental2 1.0.0-beta-2 and jsinterop-base 1.0.0-beta-3

+++ src/main/java/org/gwtproject/http/client/ResponseImpl.java
@@ -88,7 +88,7 @@ class ResponseImpl extends Response {
 
   @Override
   public int getStatusCode() {
-    return (int) xmlHttpRequest.status;
+    return xmlHttpRequest.status;
   }
 
   @Override
